extra_codes/backup/requirements-Yash.txt

diffusers==0.31.0
transformers==4.47.1
accelerate==1.2.1

protobuf==4.25.3
mediapipe==0.10.14

controlnet-aux==0.0.7

opencv-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4

huggingface-hub==0.27.1

matplotlib==3.9.4

scipy==1.13.1
einops==0.8.0
omegaconf==2.3.0
safetensors==0.4.5
filelock==3.16.1
packaging==24.2
requests==2.32.3
tqdm==4.67.1